* ... Hooooo boy... So there's a lot to talk about in this commit. I've begun implementing the Bullet Physics library. Right now nothing has collision, just the bare framework that gets initialized. Will remedy soon.
* Have FINALLY begun work on cleaning up renderer code and moving all irrelevant things to a better place (World)

* Implemented my first version of a data-oriented container in the SentenceManager class. Mostly based on the bitsquid blog / molecular musings blog. Only real difference is that I use a priority queue to find the next free ID value.

* Removed a bunch of old textures and other shit that was just taking up space.

* Changed a bunch of solution options and other things. Not really that good at the whole project/solution/OS interaction part of coding, so it'll probably be a pain in the ass to get this to work on any other pc right now.

+++ Engine/marchingCubesClass.cpp
@@ -560,8 +560,8 @@ static const XMFLOAT3 relativeCornerPositions[8] = {
 			waterMesh->SetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLESTRIP);
 
 			//Okay, so. I have already decided that the spacing will be 1.0f between each vertex.
-			unsigned int stepsX = ((chunk->GetStepCountX())*chunk->GetStepSizeX());
-			unsigned int stepsZ = ((chunk->GetStepCountZ())*chunk->GetStepSizeZ());
+			unsigned int stepsX = (unsigned int)((chunk->GetStepCountX())*chunk->GetStepSizeX());
+			unsigned int stepsZ = (unsigned int)((chunk->GetStepCountZ())*chunk->GetStepSizeZ());
 
 			vector<XMFLOAT3> vertices;
 			vector<unsigned int> indices;
@@ -577,7 +577,7 @@ static const XMFLOAT3 relativeCornerPositions[8] = {
 			{
 				for(float x = 0; x < stepsX; x++)
 				{
-					index = x + (z * stepsX);
+					index = (int)(x + (z * stepsX));
 
 					//Offset each position with a magic number to make sure there are no seams between the different water meshes.
 					vertices[index].x = (minPos.x + x-(x*0.0908f)); //0.0475f
@@ -591,9 +591,9 @@ static const XMFLOAT3 relativeCornerPositions[8] = {
 			int indexOffset = 0;
 			indices.resize(((stepsX * 2) + 2) * (stepsZ - 1));
 
-			for( int z = 0; z < stepsZ-1; z++ )
+			for(unsigned int z = 0; z < stepsZ-1; z++ )
 			{
-				for( int x = 0; x < stepsX; x++ )
+				for(unsigned int x = 0; x < stepsX; x++ )
 				{
 					index = x + ( z * stepsX );
 					indices[indexOffset] = index;
